Add get hp
