+++ Task-10_Browser-game/src/js/controllers/battleController.js
@@ -1,4 +1,4 @@
-import { getStatesTemplate, ArmorParts, getMaxHP } from "../rpgSystem";
+import { getStatesTemplate, ArmorParts, States } from "../rpgSystem";
 
 export function getDefaultCharacter() {
     let character = {
@@ -14,4 +14,32 @@ export function getDefaultCharacter() {
     };
     character.hp = getMaxHP(character);
     return character;
+}
+
+export function getMaxHP(character) {
+    let constitution = character.states[States.constitution].value;
+    let health = 0;
+    let shield = 0;
+    for (const part in character.armor) {
+        const level = character.armor[part];
+        health += getHealthByArmor(constitution, part, level);
+        shield += getShieldByArmor(constitution, part, level);
+    }
+
+    let hp = {
+        health: {
+            max: health,
+            current: health,
+        },
+        shield:  {
+            max: shield,
+            current: shield,
+        },
+    };
+
+    if(character.hp) {
+        hp.health.current = Math.round(character.hp.health.current * hp.health.max / character.hp.health.max);
+    }
+
+    return hp;
 }
