+++ Task-10_Browser-game/src/js/rpgSystem.js
@@ -98,40 +98,12 @@ export function getDamageByStrength(strength) {
     return result;
 }
 
-function getHealthByArmor(stateValue, armorPart, armorLevel) {
+export function getHealthByArmor(stateValue, armorPart, armorLevel) {
     let armor = ArmorPartDefence[armorPart](armorLevel);
     return stateValue * armor;
 }
 
-function getShieldByArmor(stateValue, armorPart, armorLevel) {
+export function getShieldByArmor(stateValue, armorPart, armorLevel) {
     let armor = ArmorPartDefence[armorPart](armorLevel);
     return armor;
-}
-
-export function getMaxHP(character) {
-    let constitution = character.states[States.constitution].value;
-    let health = 0;
-    let shield = 0;
-    for (const part in character.armor) {
-        const level = character.armor[part];
-        health += getHealthByArmor(constitution, part, level);
-        shield += getShieldByArmor(constitution, part, level);
-    }
-
-    let hp = {
-        health: {
-            max: health,
-            current: health,
-        },
-        shield:  {
-            max: shield,
-            current: shield,
-        },
-    };
-
-    if(character.hp) {
-        hp.health.current = Math.round(character.hp.health.current * hp.health.max / character.hp.health.max);
-    }
-
-    return hp;
 }
